rename contracts to more generic resolvers
