+++ pkg/flowkit/resolvers/contracts.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package contracts
+package resolvers
 
 import (
 	"fmt"
