+++ pkg/flowkit/resolvers/contracts_test.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package contracts_test
+package resolvers_test
 
 import (
 	"fmt"
@@ -30,7 +30,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/onflow/flow-cli/pkg/flowkit/contracts"
+	"github.com/onflow/flow-cli/pkg/flowkit/resolvers"
 )
 
 type testContract struct {
@@ -158,7 +158,7 @@ func (t testLoader) Normalize(base, relative string) string {
 	return relative
 }
 
-func contractBySource(all *contracts.Deployments, source string) *contracts.Contract {
+func contractBySource(all *resolvers.Deployments, source string) *resolvers.Contract {
 	for _, c := range all.Contracts() {
 		if c.Location() == source {
 			return c
@@ -205,7 +205,7 @@ func getTestCases() []contractTestCase {
 		{
 			name:                    "Two contracts with import cycle",
 			contracts:               []testContract{testContractE, testContractF},
-			expectedDeploymentError: &contracts.CyclicImportError{},
+			expectedDeploymentError: &resolvers.CyclicImportError{},
 		},
 		{
 			name:                    "Single contract with two imports",
@@ -225,7 +225,7 @@ func TestResolveImports(t *testing.T) {
 
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			c := contracts.NewDeployments(testLoader{}, noAliases)
+			c := resolvers.NewDeployments(testLoader{}, noAliases)
 
 			for _, contract := range testCase.contracts {
 				_, err := c.Add(
@@ -274,7 +274,7 @@ func TestContractDeploymentOrder(t *testing.T) {
 
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
-			c := contracts.NewDeployments(testLoader{}, noAliases)
+			c := resolvers.NewDeployments(testLoader{}, noAliases)
 
 			for _, contract := range testCase.contracts {
 				_, err := c.Add(
