+++ pkg/flowkit/resolvers/loader.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package contracts
+package resolvers
 
 import (
 	"github.com/onflow/flow-cli/pkg/flowkit"
